requirements.txt
asgiref==3.8.1
Django==5.1.2
django-filter==24.3
django-polymorphic==3.1.0
djangorestframework==3.15.2
djangorestframework-jsonapi==7.1.0
inflection==0.5.1
mysqlclient==2.2.5
Pygments==2.18.0
PyYAML==6.0.2
sqlparse==0.5.1
uritemplate==4.1.1
drf-spectacular==0.27.2
flake8==7.1.1
drf-spectacular-jsonapi==0.5.2
